requirements.txt
psycopg2
Django
requests
aiohttp
lxml
pytest-django